tweak fix compile warning
git-svn-id: http://vtp.googlecode.com/svn/trunk@6257 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainApps/CManager/ItemGroup.cpp
@@ -218,7 +218,7 @@ vtGeom *CreateRulers(vtFont *font, float fSize)
 		up++;
 		if (up > 2) up = 0;
 	}
-	int ticks = fSize / interval;
+	int ticks = int( fSize / interval );
 	FPoint3 p;
 	float *wide;
 	float *thin;
